# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LoyaltyPro API",
    "description": "API documentation for the LoyaltyPro loyalty management system",
    "version": "1.0.0",
    "contact": {
      "name": "LoyaltyPro Support",
      "email": "support@loyaltypro.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.loyaltypro.com/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint"
      }
    },
    "schemas": {
...
      "Member": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Member ID"
          },
          "name": {
            "type": "string",
            "description": "Member's full name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Member's email"
          },
          "phone": {
            "type": "string",
            "description": "Member's phone number"
          },
          "points": {
            "type": "number",
            "description": "Current loyalty points"
          },
          "tier": {
            "type": "string",
            "enum": ["Bronze", "Silver", "Gold", "Platinum"],
            "description": "Member tier"
          },
          "status": {
            "type": "string",
            "enum": ["active", "inactive", "suspended"],
            "description": "Member status"
          },
          "memberSince": {
            "type": "string",
            "format": "date-time",
            "description": "Member registration date"
          },
          "lastActivity": {
            "type": "string",
            "format": "date-time",
            "description": "Last activity timestamp"
          },
          "totalSpent": {
            "type": "number",
            "description": "Total amount spent"
          },
          "transactions": {
            "type": "number",
            "description": "Number of transactions"
          },
          "firstName": {
            "type": "string",
            "description": "First name"
          },
          "lastName": {
            "type": "string",
            "description": "Last name"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Date of birth"
          },
          "gender": {
            "type": "string",
            "description": "Gender"
          },
          "address": {
            "type": "object",
            "properties": {
              "street": {
                "type": "string"
              },
              "city": {
                "type": "string"
              },
              "state": {
                "type": "string"
              },
              "zipCode": {
                "type": "string"
              },
              "country": {
                "type": "string"
              }
            }
          },
          "emergencyContact": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "phone": {
                "type": "string"
              },
              "relationship": {
                "type": "string"
              }
            }
          },
          "preferences": {
            "type": "object",
            "properties": {
              "marketingEmails": {
                "type": "boolean"
              },
              "smsNotifications": {
                "type": "boolean"
              },
              "newsletter": {
                "type": "boolean"
              }
            }
          }
        },
        "required": ["id", "name", "email", "points", "tier", "status"]
      },
      "Analytics": {
        "type": "object",
        "properties": {
          "totalMembers": {
            "type": "number",
            "description": "Total number of members"
          },
          "activePartners": {
            "type": "number",
            "description": "Number of active partners"
          },
          "pointsCirculating": {
            "type": "number",
            "description": "Total points in circulation"
          },
          "systemRevenue": {
            "type": "number",
            "description": "Total system revenue"
          },
          "monthlyGrowth": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "month": {
                  "type": "string"
                },
                "members": {
                  "type": "number"
                },
                "revenue": {
                  "type": "number"
                }
              }
            }
          }
        }
      },
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Member"
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of items"
          },
          "page": {
            "type": "number",
            "description": "Current page number"
          },
          "perPage": {
            "type": "number",
            "description": "Items per page"
          },
          "totalPages": {
            "type": "number",
            "description": "Total number of pages"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "string",
            "description": "Detailed error information (development only)"
          }
        }
      }
    }
  },
  "paths": {
...
    "/admin/members": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get members list",
        "description": "Get paginated list of members with optional search",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "description": "Items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search term",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Members list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "Create new member",
        "description": "Create a new member in the system",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "required": true
                  },
                  "lastName": {
                    "type": "string",
                    "required": true
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "required": true
                  },
                  "phone": {
                    "type": "string"
                  },
                  "dateOfBirth": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "gender": {
                    "type": "string"
                  },
                  "address": {
                    "type": "object",
                    "properties": {
                      "street": {
                        "type": "string"
                      },
                      "city": {
                        "type": "string"
                      },
                      "state": {
                        "type": "string"
                      },
                      "zipCode": {
                        "type": "string"
                      },
                      "country": {
                        "type": "string"
                      }
                    }
                  },
                  "emergencyContact": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "phone": {
                        "type": "string"
                      },
                      "relationship": {
                        "type": "string"
                      }
                    }
                  },
                  "preferences": {
                    "type": "object",
                    "properties": {
                      "marketingEmails": {
                        "type": "boolean"
                      },
                      "smsNotifications": {
                        "type": "boolean"
                      },
                      "newsletter": {
                        "type": "boolean"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Member created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "member": {
                      "$ref": "#/components/schemas/Member"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/admin/members/{id}": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get member by ID",
        "description": "Get detailed information about a specific member",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Member ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Member"
                }
              }
            }
          },
          "404": {
            "description": "Member not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "Update member",
        "description": "Update member information",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Member ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": ["active", "inactive", "suspended"]
                  },
                  "points": {
                    "type": "number"
                  },
                  "tier": {
                    "type": "string",
                    "enum": ["Bronze", "Silver", "Gold", "Platinum"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Member updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Member"
                }
              }
            }
          },
          "404": {
            "description": "Member not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/admin/analytics": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get system analytics",
        "description": "Get comprehensive system analytics and statistics",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Analytics data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Analytics"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/admin/analytics/customers": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get customer statistics",
        "description": "Get customer count and basic statistics",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Customer statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "number",
                      "description": "Total number of customers"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
} 